+++ wayflowcore/src/wayflowcore/serialization/_builtins_components.py
@@ -173,6 +173,8 @@ _BUILTIN_COMPONENTS = {
     "ToolFromToolBox",
     "TemplateRenderingStep",
     "TextContent",
+    "TextTokenLogProb",
+    "TextTokenTopLogProb",
     "TlsOracleDatabaseConnectionConfig",
     "TlsPostgresDatabaseConnectionConfig",
     "TokenUsage",